third_party/WebKit/LayoutTests/dom/node/isEqualNode-expected.txt

Test the isEqualNode API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test isEqualNode for DocumentType nodes.
PASS docTypeAllSet.isEqualNode(docTypeAllSet2) is true
PASS docTypeAllSet.isEqualNode(docTypeDifferentPublicID) is false
PASS docTypeAllSet.isEqualNode(docTypeDifferentSystemID) is false
PASS successfullyParsed is true

TEST COMPLETE
